requirements.txt
Pillow
requests
flask
